Fold PadPathService back into PathNormalizer
Closes #64.

PadPathService was a 49-LOC wrapper around PathNormalizer that
delegated normalizeViewerFilePath unchanged and added a '.pad'
suffix in normalizeCreatePath / normalizeCreateFileName. Five
services injected it purely for that suffix.

Moved normalizeCreatePath and normalizeCreateFileName onto
PathNormalizer, deleted PadPathService and its test file. Callers
switch the constructor type / use-statement; the field name
$padPaths is preserved so call sites read identically. The two
migrated tests now live in PathNormalizerTest.

One less indirection in the DI graph. 396 tests stay green.
